Modifications and New Files
The "glycoOnto_create.py" file instantiates glycogenes as individuals in
our "glycoOnto.owl" ontology, and creates the
"glycoOnto_individuals.rdf" file.  This new RDF file is exactly the same
as "glycoOnto.owl", with the addition of all glycogenes in the "done"
table as individuals.  All glycogenes have an "is_a" relationship to the
appropriate glycogene function and Pathway classes.

The "finishedGlycogenes.tsv" is downloaded from the Google drive and
should be updated in this repository any time we update the glycogene
list.

Some class names were modified to ommit conflicting characters in the
ontology file (writing issues, object creation issues).  Such as
"Putative/Inactive Enzymes" changing to "Putative or Inactive Enzymes".