Lexing Class
There is a new lexing class which is used to parse the rules into their
attribute classes.  This class is also invoked within superclasses to
recursively identify components.

This allows for more complete parsing of rule strings and makes it
easier to detect unexpected entities within sub attributes of a rule
string.