Ontologically defined Glycans
I parsed glycans from GlyGen using a sparql query in
"glygen_glycan_query.py", and created classes similarly as was defined
in Frederique Lisacek's paper (https://doi.org/10.1371/journal.pone.0144578)

This commit has code which works in parsing these glycans from GlycoCT
and instantiating them in "glycoStructOnto", a structure ontology.

These glycans should all be human-derived, have no unknown
monosaccharide and substituent chemistry, and have no alternative or
repeating structures (based on glycoCT).  Each new instance of a
monosaccharide in a glycan is an instance in the ontology.  These
instances are linked together based on the structure of the glycan they
are associated with, based on the "has_residue" relationship.